refactor: move bulk tests into integration/crud and separate prose test
