+++ test/integration/crud/bulk.test.js
@@ -4,11 +4,10 @@ const {
   withClientV2,
   withMonitoredClient,
   setupDatabase,
-  ignoreNsNotFound
-} = require('./shared');
-const test = require('./shared').assert;
-const { MongoDriverError, MongoBatchReExecutionError } = require('../../src/error');
-const { Long, MongoBulkWriteError } = require('../../src');
+  ignoreNsNotFound,
+  assert: test
+} = require('../shared');
+const { Long, MongoBatchReExecutionError, MongoDriverError } = require('../../../src');
 const crypto = require('crypto');
 const chai = require('chai');
 const expect = chai.expect;
@@ -19,48 +18,6 @@ const MAX_BSON_SIZE = 16777216;
 describe('Bulk', function () {
   before(function () {
     return setupDatabase(this.configuration);
-  });
-
-  describe('Write Errors', () => {
-    describe('errInfo property on insertMany', () => {
-      let client;
-
-      beforeEach(async function () {
-        client = this.configuration.newClient({ monitorCommands: true });
-        await client.connect();
-      });
-
-      afterEach(async () => {
-        if (client) {
-          await client.close();
-        }
-      });
-
-      it('should be accessible', {
-        metadata: { requires: { mongodb: '>=5.0.0' } },
-        async test() {
-          try {
-            await client.db().collection('wc_details').drop();
-          } catch {
-            // don't care
-          }
-
-          const collection = await client
-            .db()
-            .createCollection('wc_details', { validator: { x: { $type: 'string' } } });
-
-          try {
-            await collection.insertMany([{ x: /not a string/ }]);
-            expect.fail('The insert should fail the validation that x must be a string');
-          } catch (error) {
-            expect(error).to.be.instanceOf(MongoBulkWriteError);
-            expect(error).to.have.property('code', 121);
-            expect(error).to.have.property('writeErrors').that.is.an('array');
-            expect(error.writeErrors[0]).to.have.property('errInfo').that.is.an('object');
-          }
-        }
-      });
-    });
   });
 
   it('should correctly handle ordered single batch api write command error', {
